refactor: move profile popout provider
